backend/jobscraper/requirements.txt
scrapy
psycopg2-binary==2.9.1